developing manager menu
